backend-server/ai-bot/requirements.txt
python-telegram-bot>=21.6
influxdb-client[async]>=1.44
openai>=1.50
python-dotenv>=1.0
